requirements.txt
transformers==4.34.1
accelerate==0.24.0
mosaicml-streaming==0.6.1
datasets
evaluate
wandb
opencv-python
scikit-learn 
scipy
pillow
pybind11
flash-attn==2.5.5
